Added function to make sure new markers are visible/hidden depending on zoom
